BACKPORT: FROMLIST: at24: Support probing while off
In certain use cases (where the chip is part of a camera module, and the
camera module is wired together with a camera privacy LED), powering on
the device during probe is undesirable. Add support for the at24 to
execute probe while being powered off. For this to happen, a hint in form
of a device property is required from the firmware.
(am from https://patchwork.kernel.org/patch/11752743/)

BUG=b:147689556
TEST=Sanity checked basic camera functions

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>

Change-Id: Ic2d4d3cca0d7ed1445562a3ac22e3357d595d58c
Signed-off-by: Bingbu Cao <bingbu.cao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2403385
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Commit-Queue: Sergey Senozhatsky <senozhatsky@chromium.org>
Tested-by: Sergey Senozhatsky <senozhatsky@chromium.org>