BACKPORT: drm/msm/dsi: add support for 7nm DSI PHY/PLL
This adds support for the 7nm ("V4") DSI PHY/PLL for sm8150 and sm8250.

Implementation is based on 10nm driver, but updated based on the downstream
7nm driver.

Signed-off-by: Jonathan Marek <jonathan@marek.ca>
Tested-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org> (SM8250)
Signed-off-by: Rob Clark <robdclark@chromium.org>

(cherry picked from commit 1ef7c99d145c2759308e53bf19f2cc971677680c)

Conflicts:
   drivers/gpu/drm/msm/dsi/phy/dsi_phy_7nm.c
   drivers/gpu/drm/msm/dsi/pll/dsi_pll.c
   drivers/gpu/drm/msm/dsi/pll/dsi_pll.h
   drivers/gpu/drm/msm/dsi/pll/dsi_pll_7nm.c

We have CL:2424532 but not the DT bits. That's OK though because we'll
remove this file shortly in a FROMGIT commit.

BUG=b:173771162
TEST=emerge-trogdor chromeos-kernel-5_4

Change-Id: I39c21b81e6f92c5c75dc98f8e5b5732eb4d89d8f
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2921275
Reviewed-by: Douglas Anderson <dianders@chromium.org>