FROMGIT: drm/msm: Fix premature purging of BO
In the case where we have a back-to-back submission that shares the same
BO, this BO will be prematurely moved to inactive_list while retiring the
first submit. But it will be still part of the second submit which is
being processed by the GPU. Now, if the shrinker happens to be triggered at
this point, it will result in a premature purging of this BO.

To fix this, we need to refcount BO while doing submit and retire. Then,
it should be moved to inactive list when this refcount becomes 0.

Signed-off-by: Akhil P Oommen <akhilpo@codeaurora.org>
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 9d8baa2bf224f6f010644b2a60a3ca73d829663e
 https://gitlab.freedesktop.org/drm/msm.git msm-next)

BUG=b:153039340
TEST=build and boot lazor

Signed-off-by: Rob Clark <robdclark@chromium.org>
Change-Id: I48fe43f5605e259d92cc182581c37c69d554519f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2424537
Commit-Queue: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>