UPSTREAM: PM: domains: Allow to abort power off when no ->power_off() callback
In genpd_power_off() we may decide to abort the power off of the PM domain,
even beyond the point when the governor would accept it. The abort is done
if it turns out that a child domain has been requested to be powered on,
which means it's waiting for the lock of the parent to be released.

However, the abort is currently only considered if the genpd in question
has a ->power_off() callback assigned. This is unnecessary limiting,
especially if the genpd would have a parent of its own. Let's remove the
limitation and make the behaviour consistent.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
[ rjw: Subject edit ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit f63816e43d90442684cd2fd74f602cf3c5fae69c)

BUG=b:177955525
TEST=CQ

Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Change-Id: I58fbec1791d95876bbca7eb6ac8992998018e44a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2636831
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>