UPSTREAM: net: ipa: fix two mild warnings
Fix two spots where a variable "channel_id" is unnecessarily
redefined inside loops in "gsi.c".  This is warned about if
"W=2" is added to the build command.

Note that this problem is harmless, so there's no need to backport
it as a bugfix.

Remove a comment in gsi_init() about waking the system; the GSI
interrupt does not wake the system any more.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 993cac15d1194b7ff03da67993b82c4b21680739
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:74020136
TEST=Booted trogdor, modem call functional

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: Ieb747b34fe01f9fe55881481d5fa4bd947656226
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2491276
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>