UPSTREAM: x86: Support Generic Initiator only proximity domains
In common with memoryless domains only register GI domains
if the proximity node is not online. If a domain is already
a memory containing domain, or a memoryless domain there is
nothing to do just because it also contains a Generic Initiator.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Acked-by: Borislav Petkov <bp@suse.de>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry-picked from commit 73bf7382debb1a93fef5ff38222c6a3b62dfea44)

BUG=b:179702302
TEST=No regressions with boot to login

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I7840f6bf2259984ad3d16f9f4f9d02f2a076e929
Signed-off-by: Chintan Patel <chintan.m.patel@intel.corp-partner.google.com>
Signed-off-by: Daniel Campello <campello@chromium.org>