UPSTREAM: thunderbolt: Add uaccess dependency to debugfs interface
Some calls in the debugfs interface are made to the linux/uaccess.h header,
but the header is not referenced. So, for x86_64 architectures, this
dependency seems to be pulled in elsewhere, which leads to a successful
compilation. However, on arm/arm64 architectures, it was found to error out
on implicit declarations.

This change fixes the implicit declaration error by adding the
linux/uaccess.h header.

Fixes: 54e418106c76 ("thunderbolt: Add debugfs interface")
Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 77455129fb5b2a8749330b2b40d0c8750b6bf076)

BUG=b:170413196
TEST=No regressions with TBT devices.

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: Ibe800275744b970310f211a8e2f4395cf2924daf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2532892
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Tested-by: Hou-hsun Lee <hou-hsun.lee@intel.corp-partner.google.com>
Commit-Queue: Prashant Malani <pmalani@chromium.org>