CHROMIUM: thunderbolt: Add uaccess dependency to debugfs interface
Some calls in the debugfs interface are made to the linux/uaccess.h header,
but the header is not referenced. So, for x86_64 architectures, this
dependency seems to be pulled in elsewhere, which leads to a successful
compilation. However, on arm/arm64 architectures, it was found to error out
on implicit declarations.

This change fixes the implicit declaration error by adding the
linux/uaccess.h header.

Fixes: 54e418106c76 ("thunderbolt: Add debugfs interface")
Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>

BUG=b:161731926, b:170413196
TEST=Build arm64:allmodconfig

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I0141f0df925a91ff0cb259f819c05297f566ad21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2462524
Tested-by: Madhulika Dolke <madhulika.dolke@intel.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>