CHROMIUM: thunderbolt: Implement sbtx and router offline port operations
To support USB4 retimer firmware update operations when no devices
are connected, port operations such as putting the router in offline
mode and set / unset of inbound side band transfer from retimer to
the SoC are required.

BUG=b:175045931
TEST=Ensure there are no regressions in using a Type-C port.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: I73bf7535259d1b216c93b7c9a49221d62ffd6199
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2639105
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Tested-by: Kranthi Kuntala <kranthi.kuntala@intel.corp-partner.google.com>