UPSTREAM: iommu/arm-smmu-qcom: Read back stream mappings
The Qualcomm boot loader configures stream mapping for the peripherals
that it accesses and in particular it sets up the stream mapping for the
display controller to be allowed to scan out a splash screen or EFI
framebuffer.

Read back the stream mappings during initialization and make the
arm-smmu driver maintain the streams in bypass mode.

Cc: <stable@vger.kernel.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Tested-by: Steev Klimaszewski <steev@kali.org>
Acked-by: Robin Murphy <robin.murphy@arm.com>
Link: https://lore.kernel.org/r/20201019182323.3162386-3-bjorn.andersson@linaro.org
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit 07a7f2caaa5a2619934491bab3c47b261c554fb0)

BUG=b:170301044
TEST=Use GPU

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I0e973182faa8fd79debf7e347e1a9c62fd92af1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2641506
Reviewed-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>