CHROMIUM: iwl7000: mvm: set enabled in the PPAG command properly
When version 2 of the PER_PLATFORM_ANT_GAIN_CMD was implemented, we
started copying the values from the command that we have stored into a
local instance.  But we accidentally forgot to copy the enabled flag,
so in practice PPAG is never really enabled.  Fix this by copying the
flag from our stored data a we should.

Change-Id: I0e8abc2b8747508b6118242533d68c856ca6dffb
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 41731f6f8f6b3f00708856492ef5cdad606ecc7c